+++ next.config.ts
@@ -1,6 +1,7 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
+
   /* config options here */
   images: {
     remotePatterns: [
@@ -9,6 +10,9 @@ const nextConfig: NextConfig = {
       { protocol: "https", hostname: "firebasestorage.googleapis.com" },
     ],
   },
+  logging:{
+    browserToTerminal: true,
+  }
 };
 
 export default nextConfig;